requirements.txt
Flask==3.0.0
Flask-WTF==1.1.1
python-dotenv==1.0.0
flask-cors==4.0.0
pymongo==4.5.0
gunicorn
flask_login
markdown
flask-limiter
